requirements.txt
Django==2.2.1
django-autoslug==1.9.4
django-tinymce4-lite==1.7.5
entrypoints==0.3
flake8==3.7.7
isort==4.3.20
jsmin==2.2.2
mccabe==0.6.1
Pillow==6.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytils==0.3
pytz==2019.1
sqlparse==0.3.0
